Update github-pages from 204 to 214
This addresses a security vulnerability in kramdown.

I also specified the Ruby version in the Gemfile, and in a .ruby-version
file. That ensures that everyone is using the same version of Ruby.

Why 2.7.2 and not the latest 3.0.0? Because Jekyll is not yet fully
compatible with 3.0.0.

Eventually, I’d like us to get rid of the `github-pages` gem. It’s not
needed to deploy to GitHub Pages, and it’s preventing us from using the
latest version of Jekyll.

Here’s a step-by-step guide I wrote for deploying to GitHub Pages using
GitHub Actions:

https://www.moncefbelyamani.com/making-github-pages-work-with-latest-jekyll/